refactor(iotdevice): rename command contract device-command → devicecommand (no-dash) [#1580]
The command contract domain `device-command` (dashed) violated the project's
no-dash identifier policy (docs/architecture/naming-guard.yaml) and was
inconsistent with its own slice `devicecommand` (concat). PR-1 renamed the slice
but left the contract dashed; DOC-NAME-01 only surfaced it now because the new
consumer guide is the first active doc to reference the contract id.

Renames all 5 command contracts (enqueue + the 4 codegen:false siblings
ack/dequeue/report/extend-lease) device-command → devicecommand: contract dirs +
ids (command.devicecommand.X.v1) + schema titles, the generated package, slice.yaml
contractUsages/verify entries, cell.go/command_handler.go/run-path imports, the e2e
+ contract tests, the consumer guide, and godoc/test contract-id examples
(runtime/command, contractgen). Regenerated boundary.yaml + types_gen + command_gen.

Siblings stay codegen:false (deferred) — only their ids are renamed, not implemented.
Descriptive Go error-prefixes ("device-command: ...") and the synthetic
domainLastSegment dash-derivation test keep the hyphenated form intentionally
(prose/derivation-coverage, not identifiers).
