+++ examples/iotdevice/cells/devicecell/slices/devicecommand/slice.yaml
@@ -13,15 +13,15 @@ contractUsages:
     role: serve
   - contract: http.device.command.extend-lease.v1
     role: serve
-  - contract: command.device-command.enqueue.v1
+  - contract: command.devicecommand.enqueue.v1
     role: handle
-  - contract: command.device-command.dequeue.v1
+  - contract: command.devicecommand.dequeue.v1
     role: handle
-  - contract: command.device-command.report.v1
+  - contract: command.devicecommand.report.v1
     role: handle
-  - contract: command.device-command.ack.v1
+  - contract: command.devicecommand.ack.v1
     role: handle
-  - contract: command.device-command.extend-lease.v1
+  - contract: command.devicecommand.extend-lease.v1
     role: handle
 verify:
   unit:
@@ -32,11 +32,11 @@ verify:
     - contract.http.device.command.report.v1.serve
     - contract.http.device.command.ack.v1.serve
     - contract.http.device.command.extend-lease.v1.serve
-    - contract.command.device-command.enqueue.v1.handle
-    - contract.command.device-command.dequeue.v1.handle
-    - contract.command.device-command.report.v1.handle
-    - contract.command.device-command.ack.v1.handle
-    - contract.command.device-command.extend-lease.v1.handle
+    - contract.command.devicecommand.enqueue.v1.handle
+    - contract.command.devicecommand.dequeue.v1.handle
+    - contract.command.devicecommand.report.v1.handle
+    - contract.command.devicecommand.ack.v1.handle
+    - contract.command.devicecommand.extend-lease.v1.handle
 
 allowedFiles:
   - examples/iotdevice/cells/devicecell/slices/devicecommand/**
