+++ examples/iotdevice/cells/devicecell/slices/devicecommand/slice_gen.go
@@ -22,11 +22,11 @@ var sliceMeta = &metadata.SliceMeta{
 		{Contract: "http.device.command.report.v1", Role: "serve"},
 		{Contract: "http.device.command.ack.v1", Role: "serve"},
 		{Contract: "http.device.command.extend-lease.v1", Role: "serve"},
-		{Contract: "command.device-command.enqueue.v1", Role: "handle"},
-		{Contract: "command.device-command.dequeue.v1", Role: "handle"},
-		{Contract: "command.device-command.report.v1", Role: "handle"},
-		{Contract: "command.device-command.ack.v1", Role: "handle"},
-		{Contract: "command.device-command.extend-lease.v1", Role: "handle"},
+		{Contract: "command.devicecommand.enqueue.v1", Role: "handle"},
+		{Contract: "command.devicecommand.dequeue.v1", Role: "handle"},
+		{Contract: "command.devicecommand.report.v1", Role: "handle"},
+		{Contract: "command.devicecommand.ack.v1", Role: "handle"},
+		{Contract: "command.devicecommand.extend-lease.v1", Role: "handle"},
 	},
 	Verify: metadata.SliceVerifyMeta{
 		Unit: []string{
@@ -38,11 +38,11 @@ var sliceMeta = &metadata.SliceMeta{
 			"contract.http.device.command.report.v1.serve",
 			"contract.http.device.command.ack.v1.serve",
 			"contract.http.device.command.extend-lease.v1.serve",
-			"contract.command.device-command.enqueue.v1.handle",
-			"contract.command.device-command.dequeue.v1.handle",
-			"contract.command.device-command.report.v1.handle",
-			"contract.command.device-command.ack.v1.handle",
-			"contract.command.device-command.extend-lease.v1.handle",
+			"contract.command.devicecommand.enqueue.v1.handle",
+			"contract.command.devicecommand.dequeue.v1.handle",
+			"contract.command.devicecommand.report.v1.handle",
+			"contract.command.devicecommand.ack.v1.handle",
+			"contract.command.devicecommand.extend-lease.v1.handle",
 		},
 	},
 	AllowedFiles: []string{
