+++ examples/iotdevice/contracts/command/devicecommand/ack/v1/contract.yaml
@@ -1,4 +1,4 @@
-id: command.device-command.dequeue.v1
+id: command.devicecommand.ack.v1
 kind: command
 codegen: false  # K#09 funnel: kind=command deferred from K#06 PR-4 (see backlog CONTRACT-YAML-CODEGEN-DEFAULT-CLEANUP @ PR559)
 ownerCell: devicecell
