+++ examples/iotdevice/contracts/command/devicecommand/ack/v1/request.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.ack.v1.request",
+  "title": "command.devicecommand.ack.v1.request",
   "type": "object",
   "properties": {
     "reason": {
