+++ examples/iotdevice/contracts/command/devicecommand/ack/v1/response.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.enqueue.v1.response",
+  "title": "command.devicecommand.ack.v1.response",
   "type": "object",
   "properties": {
     "data": {
