+++ examples/iotdevice/contracts/command/devicecommand/dequeue/v1/request.schema.json
@@ -1,5 +1,5 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.dequeue.v1.request",
+  "title": "command.devicecommand.dequeue.v1.request",
   "type": "object"
 }
