+++ examples/iotdevice/contracts/command/devicecommand/enqueue/v1/request.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.enqueue.v1.request",
+  "title": "command.devicecommand.enqueue.v1.request",
   "type": "object",
   "properties": {
     "deviceId": { "type": "string", "minLength": 1, "maxLength": 256 },
