+++ examples/iotdevice/contracts/command/devicecommand/extend-lease/v1/contract.yaml
@@ -1,4 +1,4 @@
-id: command.device-command.extend-lease.v1
+id: command.devicecommand.extend-lease.v1
 kind: command
 codegen: false  # K#09 funnel: kind=command deferred from K#06 PR-4 (see backlog CONTRACT-YAML-CODEGEN-DEFAULT-CLEANUP @ PR559)
 ownerCell: devicecell
