+++ examples/iotdevice/contracts/command/devicecommand/extend-lease/v1/request.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.extend-lease.v1.request",
+  "title": "command.devicecommand.extend-lease.v1.request",
   "type": "object",
   "properties": {
     "extensionSeconds": {
