+++ examples/iotdevice/contracts/command/devicecommand/report/v1/response.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "command.device-command.report.v1.response",
+  "title": "command.devicecommand.report.v1.response",
   "type": "object",
   "properties": {
     "data": {
