+++ tools/archtest/command_dispatch_funnel_test.go
@@ -44,7 +44,7 @@ const commandRegistryTypeName = "Registry"
 //
 // Batches A–D landed a command.Registry with RegisterHandler / LookupHandler
 // methods and a REAL generated package
-// (generated/contracts/command/device-command/enqueue/v1/command_gen.go)
+// (generated/contracts/command/devicecommand/enqueue/v1/command_gen.go)
 // that calls both methods. The generated Register and Dispatch functions are
 // the sole sanctioned callsites; a hand-written cell or example calling
 // reg.RegisterHandler(...) directly bypasses the typed Handler interface and
@@ -164,7 +164,7 @@ func TestCommandDispatchRegisterCaller01(t *testing.T) {
 	// Production scope) to confirm the sanctioned caller actually calls the guarded
 	// methods. Production() excludes generated/ by design, so we use Typed() with
 	// the generated/contracts/command/** glob (#1580: was a single hardcoded
-	// device-command/enqueue/v1 package; generalized so a second codegen command
+	// devicecommand/enqueue/v1 package; generalized so a second codegen command
 	// auto-joins coverage and removing only one funnel never silently re-vacuates
 	// this check). At least one generated package must contain a guarded call;
 	// none means the scanner regressed or every generated funnel was
