+++ tools/archtest/reverse_coverage_invariants_test.go
@@ -1030,7 +1030,7 @@ func outboxEmit(ctx context.Context, e interface{}, topic string, p interface{})
 //
 // Before #1580 the command branch only checked ownerCell != "" — a codegen:true
 // command contract could emit a typed Handler that no cell implemented and still
-// pass (the dead-but-compiles state #1580 fixes for command.device-command.enqueue.v1).
+// pass (the dead-but-compiles state #1580 fixes for command.devicecommand.enqueue.v1).
 // The command branch now additionally requires, for codegen:true commands (those
 // with a generated command_gen.go, located via loadGeneratedSourceMap(…,"command",
 // "/command_gen.go")), that ≥1 cell/example type implements the generated Handler
@@ -1066,7 +1066,7 @@ func outboxEmit(ctx context.Context, e interface{}, topic string, p interface{})
 //     caught by the anti-vacuity guard (codegen command source map non-empty ⟹
 //     ≥1 generated Handler iface must load).
 //   - Positive command-impl detection (the real types.Implements path) is exercised
-//     by THIS test's main assertion: command.device-command.enqueue.v1 is active +
+//     by THIS test's main assertion: command.devicecommand.enqueue.v1 is active +
 //     codegen:true, so if types.Implements failed to match the satellite-module
 //     EnqueueCommandAdapter, the command branch would flag it and the test would FAIL.
 //     A loader regression to root-only (GOWORK=off, dropping examples/* satellites
