+++ tools/codegen/contractgen/spec.go
@@ -431,7 +431,7 @@ type GRPCEndpointSpec struct {
 // ADR (#1044).
 type CommandSpec struct {
 	// DispatchID is the contract id used as the command.Registry key const,
-	// e.g. "command.device-command.enqueue.v1".
+	// e.g. "command.devicecommand.enqueue.v1".
 	DispatchID string
 	// HandlerMethod is the method name on the generated Handler interface:
 	// "Handle" + goPascalCase(domainLastSegment(id)).
